+++ src/data/tagData.json
@@ -1,38 +1,38 @@
 [
   {
-    "workout": "헬스",
+    "hashtag": "헬스",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   },
   {
-    "workout": "러닝",
+    "hashtag": "러닝",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   },
   {
-    "workout": "필라테스",
+    "hashtag": "필라테스",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   },
   {
-    "workout": "테니스",
+    "hashtag": "테니스",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   },
   {
-    "workout": "복싱",
+    "hashtag": "복싱",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   },
   {
-    "workout": "등산",
+    "hashtag": "등산",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   },
   {
-    "workout": "스쿼시",
+    "hashtag": "스쿼시",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   },
   {
-    "workout": "홈트",
+    "hashtag": "홈트",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   },
   {
-    "workout": "크로스핏",
+    "hashtag": "크로스핏",
     "image": "https://health.chosun.com/site/data/img_dir/2023/01/05/2023010502072_0.jpg"
   }
 ]
