+++ src/fb/myfirebase.js
@@ -1,8 +1,7 @@
 // Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
-import { getFirestore } from "firebase/firestore"; // 추가
+import { getFirestore } from "firebase/firestore";
 import { getAuth } from "firebase/auth";
-import { getDatabase } from "firebase/database"; // 확인요망
 
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
@@ -21,5 +20,4 @@ const firebaseConfig = {
 export const app = initializeApp(firebaseConfig);
 
 export const AUTH = getAuth(app);
-export const DB = getDatabase(app);
-export const POST_DB = getFirestore(app); // 추가
+export const DB = getFirestore(app);
